Moved PTOutlinedTextField.kt to core
