+++ core/build.gradle
@@ -29,6 +29,12 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+    buildFeatures {
+        compose true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion compose_version
+    }
     namespace 'tech.gusgol.core'
 }
 
@@ -38,7 +44,6 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.4.1'
 
     def room_version = "2.4.2"
-
     api "androidx.room:room-runtime:$room_version"
     api "androidx.room:room-ktx:$room_version"
     annotationProcessor "androidx.room:room-compiler:$room_version"
@@ -50,6 +55,12 @@ dependencies {
     kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
     api "androidx.hilt:hilt-navigation-compose:1.0.0"
 
+    implementation 'androidx.core:core-ktx:1.7.0'
+    implementation "androidx.compose.ui:ui:$compose_version"
+    implementation "androidx.compose.material:material:$compose_version"
+    implementation "androidx.compose.material:material-icons-extended:$compose_version"
+    implementation "androidx.compose.material3:material3:1.0.0-alpha10"
+    implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
